requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework==3.17.1
sqlparse==0.5.5
tzdata==2026.1
whitenoise==6.12.0
